+++ bidscoin/cli/_bidscoiner.py
@@ -27,13 +27,12 @@ def get_parser():
                                      epilog='examples:\n'
                                             '  bidscoiner myproject/raw myproject/bids\n'
                                             '  bidscoiner -f myproject/raw myproject/bids -p sub-009 sub-030\n'
-                                            '  bidscoiner -f myproject/raw myproject/bids -c -n "--time=00:30:00 --mem=4000"\n ')
+                                            '  bidscoiner -f myproject/raw myproject/bids -c "--time=00:30:00 --mem=4000"\n ')
     parser.add_argument('sourcefolder',       help='The study root folder containing the raw source data')
     parser.add_argument('bidsfolder',         help='The destination/output folder with the bids data')
-    parser.add_argument('-p','--participant', help='Space separated list of selected sub-# names/folders to be processed (the sub-prefix can be omitted). Otherwise all subjects in the sourcefolder will be processed', nargs='+')
-    parser.add_argument('-b','--bidsmap',     help='The study bidsmap file with the mapping heuristics. If the bidsmap filename is just the basename (i.e. no "/" in the name) then it is assumed to be located in the current directory or in bidsfolder/code/bidscoin. Default: bidsmap.yaml', default='bidsmap.yaml')
+    parser.add_argument('-p','--participant', help='Space separated list of selected sub-# names/folders to be processed (the sub-prefix can be omitted). Otherwise all subjects in the sourcefolder will be processed', metavar='LABEL', nargs='+')
+    parser.add_argument('-b','--bidsmap',     help='The study bidsmap file with the mapping heuristics. If the bidsmap filename is just the basename (i.e. no "/" in the name) then it is assumed to be located in the current directory or in bidsfolder/code/bidscoin. Default: bidsmap.yaml', metavar='NAME', default='bidsmap.yaml')
     parser.add_argument('-f','--force',       help='Process all subjects, regardless of existing subject folders in the bidsfolder. Otherwise these subject folders will be skipped', action='store_true')
-    parser.add_argument('-c','--cluster',     help='Use the DRMAA library to submit the bidscoiner jobs to a high-performance compute (HPC) cluster', action='store_true')
-    parser.add_argument('-n','--nativespec',  help='Opaque DRMAA argument with native specifications for submitting bidscoiner jobs to the HPC cluster. NB: Use quotes and include at least one space character to prevent premature parsing (default: -l walltime=00:30:00,mem=4gb)', default='-l walltime=00:30:00,mem=4gb')
+    parser.add_argument('-c','--cluster',     help='Use the DRMAA library to submit the bidscoiner jobs to a high-performance compute (HPC) cluster. You can add an opaque DRMAA argument with native specifications for your HPC resource manager (NB: Use quotes and include at least one space character to prevent premature parsing -- see examples)', metavar='SPECS', nargs='?', const='-l walltime=00:30:00,mem=4gb', type=str)
 
     return parser