+++ bidscoin/cli/_echocombine.py
@@ -26,10 +26,10 @@ def get_parser():
                                             '  echocombine myproject/bids func/*task-*echo-1* -a PAID\n ')
     parser.add_argument('bidsfolder',         help='The bids-directory with the (multi-echo) subject data')
     parser.add_argument('pattern',            help="Globlike recursive search pattern (relative to the subject/session folder) to select the first echo of the images that need to be combined, e.g. '*task-*echo-1*'")
-    parser.add_argument('-p','--participant', help='Space separated list of sub-# identifiers to be processed (the sub-prefix can be left out). If not specified then all participants will be processed', nargs='+')
-    parser.add_argument('-o','--output',      help=f"A string that determines where the output is saved. It can be the name of a BIDS datatype folder, such as 'func', or of the derivatives folder, i.e. 'derivatives'. If output = [the name of the input datatype folder] then the original echo images are replaced by one combined image. If output is left empty then the combined image is saved in the input datatype folder and the original echo images are moved to the 'extra_data' folder", default='')
+    parser.add_argument('-p','--participant', help='Space separated list of sub-# identifiers to be processed (the sub-prefix can be left out). If not specified then all participants will be processed', nargs='+', metavar='LABEL')
+    parser.add_argument('-o','--output',      help=f"A string that determines where the output is saved. It can be the name of a BIDS datatype folder, such as 'func', or of the derivatives folder, i.e. 'derivatives'. If output = [the name of the input datatype folder] then the original echo images are replaced by one combined image. If output is left empty then the combined image is saved in the input datatype folder and the original echo images are moved to the 'extra_data' folder", default='', metavar='DESTINATION')
     parser.add_argument('-a','--algorithm',   help='Combination algorithm', choices=['PAID', 'TE', 'average'], default='TE')
-    parser.add_argument('-w','--weights',     help='Weights for each echo', nargs='*', default=None, type=list)
+    parser.add_argument('-w','--weights',     help='Weights for each echo', nargs='*', default=None, type=float, metavar='WEIGHT')
     parser.add_argument('-f','--force',       help='Process all images, regardless whether target images already exist. Otherwise the echo-combination will be skipped', action='store_true')
 
     return parser