+++ bidscoin/cli/_fixmeta.py
@@ -24,7 +24,7 @@ def get_parser():
     parser.add_argument('bidsfolder',         help='The BIDS root directory that needs fixing (in place)')
     parser.add_argument('pattern',            help="Globlike recursive search pattern (relative to the subject/session folder) to select the json sidecar targets that need to be fixed, e.g. '*task-*echo-1*'")
     parser.add_argument('metadata',           help='Dictionary with key-value pairs of meta data that need to be fixed. If value is a string, then this meta data is written to the sidecars as is, but if it is a list of `old`/`new` strings, i.e. `[old1, new1, old2, new2, ..]`, then the existing meta data is re-written, with all occurrences of substring `old` replaced by `new`', type=json.loads)
-    parser.add_argument('-p','--participant', help='Space separated list of sub-# identifiers to be processed (the sub-prefix can be left out). If not specified then all participants will be processed', nargs='+')
-    parser.add_argument('-b','--bidsmap',     help='Selects a custom study bidsmap file for extracting source data properties and attributes. If the bidsmap filename is just the basename (i.e. no "/" in the name) then it is assumed to be located in the current directory or in bidsfolder/code/bidscoin. Default: bidsmap.yaml or else the template bidsmap')
+    parser.add_argument('-p','--participant', help='Space separated list of sub-# identifiers to be processed (the sub-prefix can be left out). If not specified then all participants will be processed', nargs='+', metavar='LABEL')
+    parser.add_argument('-b','--bidsmap',     help='Selects a custom study bidsmap file for extracting source data properties and attributes. If the bidsmap filename is just the basename (i.e. no "/" in the name) then it is assumed to be located in the current directory or in bidsfolder/code/bidscoin. Default: bidsmap.yaml or else the template bidsmap', metavar='NAME')
 
     return parser