+++ docs/CHANGELOG.md
@@ -8,6 +8,9 @@
 - A bidseditor context menu for comparing selected run-items
 - A bidseditor context menu for changing the data type of selected run-items
 
+### Changed
+- Merged the DRMAA `cluster` and `nativespec` input arguments
+
 ## [4.3.2] - 2024-03-29
 
 ### Fixed